refactor(nmsystem): one stem format vocabulary for the store, the writer and the encoder
StemFormats moves to NoMercy.NmSystem.Domain, beside MediaTypes: both are a
vocabulary several assemblies have to agree on, and the encoder is the third
assembly that needs this one. Its MIME table drops its own "audio/ogg",
"audio/opus" and "audio/flac" literals for the constants - the third copy of
the table goes with them, and a file the encoder labels one way and a
register row labels another is a client handed something other than what it
asked for. StemFormatsTests moves with the type; OutputArtifactTests pins
the three extensions against the vocabulary rather than against a repeat of
the same literals.
